pyproject.toml
[tool.poetry]
name = "test-task-parser"
version = "0.1.0"
description = ""
authors = ["Mikhail Orlov <crazyminer398@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
scrapy = "^2.11.2"
python-dotenv = "^1.0.1"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
